+++ backend/src/main/java/com/example/backend/dto/ErrorDto.java
@@ -1,6 +1,6 @@
 package com.example.backend.dto;
 
-import com.example.backend.common.exception.ResponseStatus;
+import com.example.backend.common.response.ResponseStatus;
 import lombok.Getter;
